# [WIP] Optimize encoding and decoding float64 slices