cc @cw9 @jeromefroe 

Need to fix the tests that have been commented out so marking it as work in progress, but the core implementation is complete. Benchmarks have shown promising results with the optimizations in this PR for encoding and decoding float64 slices, which is a significant portion of the overall encoding and decoding work.

Benchmark results (native is the current implementation, packed is the new implementation)
```
BenchmarkEncodeFloat64NativeSmall-8    	 2000000	       545 ns/op
BenchmarkEncodeFloat64NativeMedium-8   	   50000	     36431 ns/op
BenchmarkEncodeFloat64NativeLarge-8    	    1000	   1864905 ns/op
BenchmarkEncodeFloat64PackedSmall-8    	 5000000	       406 ns/op
BenchmarkEncodeFloat64PackedMedium-8   	  100000	     19447 ns/op
BenchmarkEncodeFloat64PackedLarge-8    	    2000	   1038561 ns/op
BenchmarkDecodeFloat64NativeSmall-8    	 1000000	      1066 ns/op
BenchmarkDecodeFloat64NativeMedium-8   	   20000	     65211 ns/op
BenchmarkDecodeFloat64NativeLarge-8    	     500	   3669564 ns/op
BenchmarkDecodeFloat64PackedSmall-8    	 5000000	       298 ns/op
BenchmarkDecodeFloat64PackedMedium-8   	  200000	     12088 ns/op
BenchmarkDecodeFloat64PackedLarge-8    	    2000	    658596 ns/op
```

As shown above, with the packed encoding/decoding, for medium sized `[]float64` with 1120 values, the new encoder is ~1.9x faster than the current encoder and the new decoder is ~5.4x faster than the current decoder.